requirements.txt
torch==1.12.1+cu113
torchvision==0.13.1+cu113
torchattacks==3.2.4
albumentations==0.5.2
hydra-core==1.2.0
omegaconf==2.2.3
contexttimer==0.3.3
tabulate==0.9.0
scikit-image==0.19.3
scikit-learn==1.1.3
Pillow==9.2.0